refactor: move interfaces to different file
